Fixed some compiler & javadoc warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@318 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ fs/src/fs/org/jnode/fs/ext2/BlockBitmap.java
@@ -49,7 +49,7 @@ public class BlockBitmap extends FSBitmap {
 		int nonfullBitmap=-1;	//points to a nonfull byte in the bitmap
 		
 		//skip the metadata (superblock copy?, groupgrescriptor copies?, inode table)
-		int first = (int)Math.ceil((double)metadataLength/8.0);
+		int first = (int)Math.ceil(metadataLength/8.0);
 		
 		for(int i=first; i<data.length; i++) {
 			if(data[i]==0x00) {
